ossl netstream driver: allow ephemeral Diffie-Hellman key exchange
Use well known DH parameters that have built-in support in OpenSSL.
From the man page: If "auto" DH parameters are switched on then
the parameters will be selected to be consistent with the size of
the key associated with the server's certificate.  If there is no
certificate (e.g. for PSK ciphersuites), then it it will be
consistent with the size of the negotiated symmetric cipher key.