build fix: regression from recently merged 615cde156c14
The clang Pragma()s were not correctly enabled due to the new
gcc version check. Also, gcc 4.2 upwards was not processed
correctly. This broke rsyslog build on the affected platforms.

Note: checking the exact version of clang is more complex than
worth the effort. Even very old versions (down to v 3) support
the macros. So we just check if we compile via clang.