tcpsrv: move worker thread pool to server instance
The worker thread pool was a past experiment. It was kept inside the
code and it "obviously" worked. However, it was now seen that the
implementation with global variable interferes with multiple
users of the tcpsrv object. Namely, it could theoretically
cause issues between imtcp and imdiag, which routinely share
this object inside the testbench.

To fix this, we have now moved the thread pool and its helper
variables to the tcpsrv object instance. We preserve the very
rough implementation for now.

In this commit are also some other changes to prep for an
improved threading model (very early stage, experimental).
However, they are currently disabled).