tcpsrv: improve multithreading *** WORK IN PROGRESS ***
Up to this commit, imtcp had no real internal multithreading
support. It includes a very rough worker pool from a past
experiment many years ago. This was never carried on. This
commit implements a fully blow worker pool, enabling imtcp
to perform much better in high-demand scenarios.

This is especially important as imtcp supports TLS and as
such can be compute-intense.

Along this we, there are also a number of related bug fixes
and enhancements, which are usually NOT provided in separate
commits, because that would complicate matters with this main
patch much more. They are listed below in the commit message.

Details:
========

* fixes to past worker thread pool experiment

The worker thread pool was a past experiment. It was kept inside the
code and it "obviously" worked. However, it was now seen that the
implementation with global variable interferes with multiple
users of the tcpsrv object. Namely, it could theoretically
cause issues between imtcp and imdiag, which routinely share
this object inside the testbench.

To fix this, we have now moved the thread pool and its helper
variables to the tcpsrv object instance. We preserve the very
rough implementation for now.

* refactor session accept to it's own function