imdiag bugfix: racy network driver selection
The testbench helper imdiag did not properly set the network stream
driver to use but instead relyed on the default one. Depending on
timing, this could mean that either ptcp or a configured default
driver could be used, e.g. one of the TLS drivers. With TLS,
processing did not happen correctly and could lead to abort.

In any case, imdiag was never meant to be used with any other
driver than ptcp (plain tcp). It just uses the tcpsrv component
in order to not implement another server.

The year-old problem is possibly the cause for some flakiness in
some of the TLS tests, for which we could not find the root cause
so far.

In any case, we now request the ptcp driver hardcoded, which not
only solves the issue but is the proper thing to do.

Note: imdiag is a testbench tool and we do not expect that someone
uses it in production.