not finished: remove epoll control structure list
We move the dynamically allocated elements to other, already
existing objects. This simplifies overall processing, both
from a speed as well as code complexity PoV.

Note: this is potentially an interim refactoring, that shall be
further refactored in the future. I am not yet fully convinced
that this is the best method. But it is better than the one
before.