step: add new module/input param "workerthreads" to imtcp
These parameters permit to specify the number of worker threads
to be used inside the worker pool. Note: if thread is one, there
is no worker pool itself, as we always have a dedicated thread running.
With workerthreads being two or more, the default thread will just
handle envent notifications and the actual communication work is
handled by the worker threads. The default currently is two.