CI: reduce parallel test runs
We are seeing an increasing number of flaky tests with hard-to-explain
failure conditions. Upon analysis it looks like this primarily stems
back to overloaded CI VMs. This can be the case because we have
performance enhanced rsyslog of the past month, which means the VMs
have higher load (due to better parallel processing) at the same time.
This is especially the case with CPU intense checks like TSAN.

In order to address this, we reduce the number of parallel tests.