tcpsrv: fix non-maintained stats counter "run"
This affects primarily imtcp. The counter was not maintained at all,
it now is incremented whenever a processing loop of the worker thread
is completed. Note that, for busy senders, can mean a large number
of iterations. The other counters tell the the details of operation.
That also means that very busy threads can be detected by looking at
the ratio of runs vs. actual operations.

This patch also fixes an invalid mutex being locked on systems that
do not support atomic instructions (a copy&paste errror).