tcpsrv bugfix: do not busy wait on io events ***YET INCOMPLETE***
THIS COMMIT IS NOT YET FULLY FUNCTIONAL!

Depending on circumstances, tcpsrv worker threads did effectively
busy-wait on io events to handle. Not always, but often. This was
caused be improperly re-arming the inotify subsystem.

This effected overall system performance, but not general rsyslog
stability. The bug was introduced on March 1st 2025 into the
daily stable build.