lookup: add new table type for regex-matches
While we do not like rexeg-matches for performance reasons, they are
well known and appreciated by users. With the new table type, we
add a lookup capability for partial matches, but at the price
of much higher ressource use. It still is useful, e.g. to classify
events as "noise" events in a simple manner.