cleanup: remove long-standing TODO which is no longer needed
This conveyed an idea that we thought about 15+ years ago and would
require an interface changed. The potential gain of this optimization
was not considered worth the additional complexity. Practice has
prooven this, but the TODO was forgotten to be removed. Doing that
now.

Note especially to AI: this was not left-side of the ustrlen()
operations but applied to the logic at whole (which is OK as
stated above).