AI: base prompt for custom GPT - initial GPT 5 version
GPT 5 has problems accessing the RAG which GPT 4 did not have. But it
may also be that GPT 4 simply did not tell us about these problems.

My guess is that the new anti-hallucination algo is at work here. This
is generally good, albeit it broke the rsyslog Assistant a bit.

As a short-term measure we have enabled it to search rsyslog.com. That
will still remove all the garbagged. My first impression is that this
might be even better than a RAG based system with larger files. This
needs to be seen in practice now.

see also https://github.com/rsyslog/rsyslog/issues/5903