ci: add clang static analyzer workflow via github action (#5981)
This replaces the former clang static analyzer check via buildbot, which needed quite some baby-sitting due to environment constraints. For the next time, run it via github action.

Note that we use the same core script to run the analyzer like we did previously. This means we can also switch back with ease.