community: add welcome message to first few commits (#6032)
intent is two-fold:
- act welcoming and provide some guidance for PR contributors
- do a bit of checking so maintainer get's an idea if this might
  be a spammy account (the world is a bad place...). Also
  prevent rsyslog from being used for social engineering other
  repos.

This method will be improved in the future and is work in progress.

Note: rsyslog welcomes small PRs, even single typo fixes, when
they are useful to the project. We shall, however, reject frequent
single typo fixes from same account and ask them to combine the
commits. Reason is this might be a social engineering tactic to
make rsyslog itself or other repos build unrooted trust in the
account.