net: Add NetworkNamespace APIs
This moves NetworkNamespace functionality into
the net module. This allows the same code to
be reused across multiple tools and plugins.
The first usage is with omfwd, which is changed
to use the common net implementation.  Note
the net implementation is based on the original
omfwd implementation.  Subsequent PRs will be
opened for integrating this into omuxsock and
imtcp.

The original test case tcp_forwarding_ns_tpl.sh
was broken due to use of single quotes rather
than double quotes, thus preventing the proper
port number argument to be passed to the
listener.  Note this test must be run as
root.

Development has occurred across Fedora41 and
Fedora42, which uses glibc 2.40 and glibc2.41
respectively.  The valgrind suppressions are
updated to handle new glibc issues
accordingly.

Four new functions are callable through the
net module. These allow one to save a handle
to the current namespace, switch to a new
namespace by name, and restore the namespace
from the saved handle.  A fourth higher
level wrapper is used to open a socket in
a named network namespace, and handles the
invocation of the lower level functions.
Ideally this would be the only public
function, however it simplifies integration
into imtcp in the future (which doesn't
directly open sockets).  This may change
in the future as network namespaces are
integrated into more modules and plugins.

Signed-off-by: Billie Alsup <balsup@cisco.com>