dev container: do not work-around libgcrypt dev interface change
there was "workaround" introduced in the container which prevented
to detect some issues in libgrypt-related code. This was probably done
to get newer version in dev env without the need to adapt code to
broken libgrypt API level in new version.

That prevented issues in https://github.com/rsyslog/rsyslog/pull/5406
to be deteced.