CI: remove outdated ElasticSearch v6.0.0 testing
This is a first step in getting the omelasticsearch testbench in better
shape. For now, the ES 6 download requirements are just commented out,
to check if that is at least a feasible measure. We also need a quick
fix to get CI running again.

Once the initial problem is solved, follow-up patches will be used to
modernize the omelasticsearch testbench.