doc: split property documentation into separate files
This is part to make the documentation more accessible and easier to read and find (for human, SEO and AI).

Also, the documentation content has been improved, e.g. by adding additional information which was previously (incorrectly) assume to be present by users.

This is a stepping stone which provides much value. But there is still lots of room for improvement.