CI: fix imtcp-tls-gibberish being executed in non-TLS builds
This tool requires rsyslog to build with TLS support. If not present,
the test will always fail. This is solved by executing it only if gnutls
is enabled. As this is a fequently tested environment, this does not
reduce test coverage. It is easier to do then checking for both gnutls
and openssl.

Many thanks to Michael Biebl for bringing this to our attention.

closes https://github.com/rsyslog/rsyslog/issues/6224