Fix: Remove stray/unwanted 'rsyslog' subproject
This was introduced in 78541ff, which I accidentally merged with the
subproject (I should have rejected merge unless this was fixed.).

Now that I notice, this is the clean-up. Sorry for any confusion this
might have caused.

Thanks for Juha Kallioinen for alerting me.

see also: https://github.com/rsyslog/rsyslog/discussions/6301