doc: add json-ld (basic schema) (#6303)
* doc: add json-ld to doc set (base schema)

This adds basic json-ld to the doc set so that articles are properly flagged as techArticle or FAQ.
This also adds a toggle switch and user doc on how to disable json-ld generation. That is expected to be done when building versions for distro packages for offline consumption.

Note that while this method improves semantic understanding, we do not expect a notable improvement in large-scale RAG systems, as the chunking/indexing usually works well enough. This is not yet graph-RAG json-ld - we need to work more on the ontology.

With the help of AI-Agent: ChatGPT