TEST: ratelimiter implementation via google antigravity
This is a know-complex request and I wanted to see how well antigravity
handles it. So far, no other AI agent was successful.

It is not clear if the generated artifact actually fully speaks for
antigravities w/ Gemini 3 Pro capabilities. We hit rate limits and
needed to switch models. Anyhow, as a first impression it for sure is
useful.