CI: build doc on pull_request, not pull_request_target
With _target, we always get outdated doc into CI, which does not
really work for testing. So we switch back to just pull_request.
In this case, we cannot deploy the doc, but that's kind of OK.

We can probably later clean up the workflow once we know this change
here actually fixes the situation.